enums: Macro definition in separate file
